requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
mysql-connector-python==8.1.0
textblob==0.17.1
requests==2.31.0
python-dotenv==1.0.0